telegram_bot/requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
httpx==0.26.0
openpyxl==3.1.2
python-dotenv==1.0.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
pydantic==2.5.2
redis==5.0.1
aiofiles==23.2.1
